Added GrpcServices

+++ ApComparison.Domain/Entities/Ingredient.cs
@@ -3,7 +3,7 @@
 public record Ingredient : BaseEntity
 {
     public required string Name { get; set; }
-    public required decimal Quantity { get; set; }
+    public required double Quantity { get; set; }
     public required string UnitOfMeasure { get; set; }
     public ICollection<Dish> DishIngredients { get; set; } = null!;
 }
