+++ ApiComparison.Contracts/ResponseDto/IngredientResponseDto.cs
@@ -3,7 +3,7 @@
 public record IngredientResponseDto : BaseResponseDto
 {
     public required string Name { get; set; }
-    public required decimal Quantity { get; set; }
+    public required double Quantity { get; set; }
     public required string UnitOfMeasure { get; set; }
     public ICollection<DishResponseDto> DishIngredients { get; set; } = null!;
 }
